Remove file ref count logic from CacheFS module. Memory usage was scaling with project size which eventually resulted in unacceptable memory usage. Also fixed a couple of warnings and errors in the test suite.
