+++ lib/cache/cache_fs.js
@@ -9,29 +9,8 @@ const moment = require('moment');
 const pick = require('lodash').pick;
 const crypto = require('crypto');
 const fileExtensions = require('lodash').invert(consts.FILE_TYPE);
-const cluster = require('cluster');
 
 class CacheFS extends CacheBase {
-    constructor() {
-        super();
-        this._id = ++CacheFS._idCounter;
-        this._kMsgIncrementGuidRef = `_incrementGuidRef.${this._id}`;
-        this._kMsgReleaseGuidRef = `_releaseGuidRef.${this._id}`;
-        this._guidRefs = {};
-
-        if(cluster.isMaster) {
-            cluster.on('message', (worker, msg) => {
-                if(!msg._msg) return;
-
-                if(msg._msg === this._kMsgIncrementGuidRef) {
-                    this._incrementGuidRef(msg.key);
-                }
-                else if(msg._msg === this._kMsgReleaseGuidRef) {
-                    this._releaseGuidRef(msg.key);
-                }
-            });
-        }
-    }
 
     get _optionsPath() {
         return super._optionsPath + ".cache_fs";
@@ -71,39 +50,6 @@ class CacheFS extends CacheBase {
 
     /**
      *
-     * @param {String} key
-     * @private
-     */
-    _incrementGuidRef(key) {
-        if(cluster.isWorker) {
-            return process.send({_msg: this._kMsgIncrementGuidRef, key: key });
-        }
-
-        if(this._guidRefs.hasOwnProperty(key)) {
-            this._guidRefs[key]++;
-        }
-        else {
-            this._guidRefs[key] = 1;
-        }
-    }
-
-    /**
-     *
-     * @param {String} key
-     * @private
-     */
-    _releaseGuidRef(key) {
-        if(cluster.isWorker) {
-            return process.send({_msg: this._kMsgReleaseGuidRef, key: key });
-        }
-
-        if(this._guidRefs.hasOwnProperty(key)) {
-            this._guidRefs[key]--;
-        }
-    }
-
-    /**
-     *
      * @param {String} type
      * @param {Buffer} guid
      * @param {Buffer} hash
@@ -127,10 +73,6 @@ class CacheFS extends CacheBase {
     async _writeFileToCache(type, guid, hash, sourcePath) {
         const filePath = this._calcFilepath(type, guid, hash);
 
-        if(this._guidRefs.hasOwnProperty(filePath) && this._guidRefs[filePath] > 0) {
-            throw new Error(`File is busy, cannot overwrite ${key}`);
-        }
-
         await fs.move(sourcePath, filePath, { overwrite: true });
         return filePath;
     }
@@ -147,13 +89,10 @@ class CacheFS extends CacheBase {
 
         return new Promise((resolve, reject) => {
             stream.on('open', () => {
-                this._incrementGuidRef(key);
                 resolve(stream);
             }).on('error', err => {
                 helpers.log(consts.LOG_ERR, err);
                 reject(err);
-            }).on('close', () => {
-                this._releaseGuidRef(key);
             });
         });
     }
@@ -300,8 +239,6 @@ class CacheFS extends CacheBase {
         this.emit('cleanup_delete_item', item.path);
     }
 }
-
-CacheFS._idCounter = 0;
 
 class PutTransactionFS extends PutTransaction {
     /**
